pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "server"
version = "0.1.0"
description = "Bias identification plugin to help readers identify and understand the bias of their media sources"
authors = ["Georgiy Sekretaryuk <37413530+sekretaryuk@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic ="^1.8.2"
fastapi = "^0.108"
uvicorn = "^0.25"
python-dotenv = "^1.0.1"
openai = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[tool.poetry.scripts]
run-model = "model.dummy_model:main"